rename composite query plan
